+++ Elements/src/Geometry/Solids/Solid.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Text;
 using Elements.Geometry.Interfaces;
+using Elements.Search;
 using Elements.Spatial;
 using LibTessDotNet.Double;
 
@@ -580,9 +581,8 @@ namespace Elements.Geometry.Solids
                     }
                 }
 
-
                 var d = facePlane.Normal.Cross(p.Normal).Unitized();
-                edgeResults.Sort(new DotComparer(d));
+                edgeResults.Sort(new DirectionComparer(d));
 
                 // Draw segments through the results and add to the 
                 // half edge graph.