+++ CHANGELOG.md
@@ -24,7 +24,10 @@
 ### Fixed
 
 - Deduplicate catalog names during code generation.
+  <<<<<<< HEAD
+- # Fix some issues with code generation and deserialization of `Vector3` and `Mesh` types.
 - Fixed an issue where GLTFs would occasionally be generated with incorrect vertex normals.
+  > > > > > > > 3a8c28c8c385007084e1160636a595d7faaaab55
 
 ## 0.9.2