+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - `Line.Intersects(BBox3 box, out List<Vector3> results, bool infinite = false)`
 - `Vector3.AreCoplanar(Vector3 a, Vector3 b, Vector3 c, Vector3 d)`
 - `Line.IsAlmostEqualTo(Line line)`
+- `ConvexHull.FromPointsInPlane(IEnumerable<Vector3> points, Vector3 normalVectorOfFrame)`
+
 ### Changed
 
 ### Fixed